Triton Server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.4
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
exceptiongroup==1.2.2
fastapi==0.115.10
filelock==3.17.0
fsspec==2025.2.0
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
httplib2==0.20.2
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
numpy==1.26.3
oauthlib==3.2.0
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-multipart==0.0.20
python-rapidjson==1.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
ssh-import-id==5.11
starlette==0.46.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
tritonclient==2.55.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wadllib==1.3.6
zipp==1.0.0
zope.event==5.0
zope.interface==7.2
